Update Move-TODOBlock.ps1. Add UtilityLib.ps1.
Polished up and worked on the Assert-ValidTODOBlocks function in the Move-TODOBlock.ps1 script. Added parameter sets and attributes to the parameters in the Move-TODOBlock.ps1 script. Reorganize utility functions into the UtilityLib.ps1 script to keep the project clean. Other small fixes and clean-ups included in the program, such as adding recursive file iteration, and updating the TODO comment block parser logic.